Update to maxSize method
hello :) you're initializing the max with 1 and that is not correct, in case with a matrix with all zeros it will give us '1' and the right answer is '0',  we must initialize it with zero and then while copying the first row and col. check if there is element is 1 then set max to 1, otherwise max will be left with 0.